+++ backend/canisters/group_index/impl/src/model/moderation_flags.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 bitflags! {
     #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
     pub struct ModerationFlags: u32 {
-        const OFFENSIVE = 0b00000001;
-        const ADULT     = 0b00000010;
-        const SPAM      = 0b00000100;
+        const OFFENSIVE     = 0b00000001;
+        const ADULT         = 0b00000010;
+        const UNDER_REVIEW  = 0b00000100;
     }
 }